Move metadata into the `Internal` module
It's not part of the public API. We could expose the type, but there's
(at the moment) nothing you would be able to do with it. Better for it
not to show up in documentation then.
